Add integration tests for drafts module (#53)
* tests(drafts_module): add integration tests

added all integration tests in the drafts
module

* test(drafts_module): update base integration test

updated the base integration test class with ClearDatabaseAsync. This is
called when the database is
initialized and when it is disposed so that all
tests use a fresh database instance. This is
important for many of the tests
that test the number of objects in the database

closes #52