feat(drafts_module): add api routes
add routes to start, pause, and complete a draft.
add route to add a pick to
a draft
add a route to remove a drafter from a draft
add get routes to get picke, draft
positions, latest drafts, and
upcoming drafts
add drafter routes to assign trivia results, execute vetoes and veto
overrides,
and get trivia results
update the tags for some routes
Fixed the AssignTriviaResultsCommandHandler, ComleteDraftCommandHandler,
and
StartDraftCommandHandler to use an updated version of the draft
from
the database.
Fixed the draft pick queries (and repository) to return the drafter name
and
movie name along with the id's of each.