refactor(drafts): refactors drafts endpoints
Refactor namespaces in Drafts module

Reorganized namespaces for classes in the `ScreenDrafts.Modules.Drafts.Presentation` module.
Classes related to draft deletion are now in the `Deletes` namespace, retrieval classes in `Gets`,
and creation and update classes in `Posts` and `Updates`, respectively.
This restructuring enhances code organization and clarity.
Additionally, made minor adjustments in the `UserRegisteredIntegrationEventConsumer` class.